requirements.txt
simplematrixbotlib==2.7.0
matrix-nio[e2e]==0.19